pulls in img client
needed to modify this codebase in order to properly release file locks after
build solve is complete. this is a temporary solution until we can re-write
this logic strictly using the buildkit client library.

also nested build func so that we can test image sizes and then finally push
an image to a OCI registry