creates containerd metadata stores during init
this was originally done inside createWorkerOpt which required the
invocation of this method to get/push/tag images. it makes more sense
for the client to hold refs to these stores and use them apart from the
workerOpt.

also renamed a couple of methods, like Push -> PushImage, so there's
consistency in the public API.