tweak build job creation; add explicit PSP/SCC support, empty dir mount for state, trapping of interrupt signals (#45)
* dump cluster info in CI to ease debugging

* run image as a numeric ID so that we don't need a custom PSP

* build job: trap signals and display a message

* remove explicit PSP creation

Namespaced resources cannot "own" global resources so we must
create it in the Helm chart and pass in the reference. This also adds
SCC support for OpenShift.

* move image build "state" into an empty directory

Can't do "overlay on overlay" so make sure files & state are stored in
an explicit emptyDir mount. Also adds support for special SELinux constraints.

* fix a TODO comment, remove ActiveDeadlineSeconds for real