fixes jobconfig vol bug
adding vols/mounts to the same field that is uses by the external config
results in a growing list of duplicates. furthermore, we cannot "blank
out" that list since it may contain user-specific vols. hence, this adds
new fields that can safely be reset after use instead of creating a more
complex strategy for using the existing vol/mount fields.